extract transformPasted
